refactor: clean up thoughts summary
